Clean up conditional definition of Coord
