+++ embedded-graphics/src/primitives/line.rs
@@ -2,6 +2,7 @@
 
 use super::super::drawable::*;
 use super::super::transform::*;
+use coord::Coord;
 
 // TODO: Impl Default so people can leave the color bit out
 /// Line primitive
@@ -147,7 +148,7 @@ impl Transform for Line {
     /// ```
     /// # use embedded_graphics::primitives::Line;
     /// # use embedded_graphics::transform::Transform;
-    /// # use embedded_graphics::drawable::Coord;
+    /// # use embedded_graphics::coord::Coord;
     ///
     /// let line = Line::new(Coord::new(5, 10), Coord::new(15, 20), 1);
     /// let moved = line.translate(Coord::new(10, 10));
@@ -168,7 +169,7 @@ impl Transform for Line {
     /// ```
     /// # use embedded_graphics::primitives::Line;
     /// # use embedded_graphics::transform::Transform;
-    /// # use embedded_graphics::drawable::Coord;
+    /// # use embedded_graphics::coord::Coord;
     ///
     /// let mut line = Line::new(Coord::new(5, 10), Coord::new(15, 20), 1);
     /// line.translate_mut(Coord::new(10, 10));
